ByPass authentication for requests redirected  internally
When If-Match header with etag value is passed for PATCH requests, new request with
GET method is internally framed and sent to calculate the current etag value and match with
provided etag. For this request by pass authentication since the same request is initially
authenticated and privileges are assigned. Also to the new request framed internally, same req.session
is assigned. So its duplicate effort in authenticating and assigning the privilages.
Also this causes the BMCWEB to crash when assigning the privilages.

Tested: Run Redfish Protocol validator and got passed.

Signed-off-by: arunthomas <arunthomasb@ami.com>

Signed-off-by: Pranoy Goru <gorumanip@ami.com>